+++ Main/index.py
@@ -1,217 +1,87 @@
 from datetime import datetime, date
 
 
+
 """
 
 
 """
 class Task:
       """
-      """
+      Task class that contains attributes including task number,title, required skills, credit hours, number of volunteers, and status.
+      """      
+      tasksRecord = []
 
-      taskRecord = {}      
       taskIdCounter = 0
       
-      def __init__(self) -> None:
+      def __init__(self, title = "", skills = "", creditHours=0, numOfVol=0):
             Task.taskIdCounter +=1
-            
-            print("Enter the following information to create your task")
+                        
             self.__taskNo = Task.taskIdCounter
-            self.__title = str(input("Task title: "))
-            self.__requiredSkills = str(input("Skills (separated by comma): ").split(","))
-            self.__creditHour = int(input("Credit Hours: "))
-            self.__numOfVolunteersNeeded = int(input("Number of Volunteers: "))
+            self.__title = title
+            self.__requiredSkills = skills.split(",")
+            self.__creditHour = creditHours
+            self.__numOfVolunteersNeeded = numOfVol
             self.__status = "Newly Added"
-            
-            
-            
-            Task.taskRecord [self.__taskNo] = {
-                  "taskNo": self.__taskNo,
-                  "title": self.__title,
-                  "required skills": self.__requiredSkills,
-                  "credit hour": self.__creditHour,
-                  "number of volunteers needed": self.__numOfVolunteersNeeded,
-                  "status": self.__status
-            }
+            Task.tasksRecord.append(self)
             
             # getters & setters
-      def setTaskNo(self,taskNo):
+      def setTaskNo(self,
+                    taskNo):
             self.__taskNo = taskNo
       
       def getTaskNo(self):
             return self.__taskNo
       
-      def setTitle(self,title):
+      def setTitle(self,
+                   title):
             self.__title = title
       
       def getTitle(self):
             return self.__title
       
-      def setRequiredSkills(self,skills):
+      def setRequiredSkills(self,
+                            skills):
             self.__requiredSkills = skills
       
       def getRequiredSkills(self):
             return self.__requiredSkills
       
-      def setCreditHour(self,hour):
+      def setCreditHour(self,
+                        hour):
             self.__creditHour = hour
       
       def getCreditHour(self):
             return self.__creditHour
       
-      def setNumOfVolunteerNeeded(self,num):
+      def setNumOfVolunteerNeeded(self,
+                                  num):
             self.__numOfVolunteersNeeded = num
+            """
+            Global functions 
+            """
+      def getNumOfVolunteerNeeded(taskNo):
+            for task in Task.tasksRecord:
+                  if task.getTaskNo() == taskNo:
+                           return f"This task requires {task.__numOfVolunteersNeeded} persons to be accomplished"
+      #Exception Handling, task number not found!!!
       
-      def getNumOfVolunteerNeeded(self):
-            return self.__numOfVolunteersNeeded
+      def updateStatus(self,status):
+            self.__status = status
+      #Exception Handling, task number not found!!!
       
-     
-      def getStatus(self, taskNo):
-            infoOfStatus = ""
-            for task, data in Task.taskRecord.items(): 
-                  if task == taskNo:
-                        infoOfStatus = data['status']
-            return infoOfStatus
-      
-      def updateStatus(self,taskno ,status):
-            if taskno in Task.taskRecord:
-                  for task, data in Task.taskRecord.items(): 
-                        if task == taskno:
-                              data['status'] = status
-                              print(f"Status has been updated for task {taskno}")
-                        else: 
-                              print("Loading..")
-            else:
-                  print(f"Id {taskno} not found in the dictionary")
-            
-
-
-
-class Volunteer_Opportunity:
-      """
-      """
-      opportunityDict = {}
-      opportunityCounter = 0
-      
-      
-      def __init__(self) -> None:
-            Volunteer_Opportunity.opportunityCounter += 1
-            self.__title = str(input("Volunteer Opportunity: "))
-            self.__date = input("date of birth (yyyy-mm-dd): ").split("-")
-            self.__date = date(int(self.__DOJ[0]), int(self.__DOJ[1]), int(self.__DOJ[2]))
-            self.__startingTime = str(input("Starting Time (00:00): "))
-            self.__endingTime = str(input("Ending Time (00:00): "))
-            self.__location = str(input("Location: "))
-            self.__tasks = []
-            self.__assignedVolunteers = []
-            self.__interest = []
-            
-            Volunteer_Opportunity.opportunityDict [Volunteer_Opportunity.opportunityCounter] = {
-                  "title":self.__title,
-                  "date":self.__date,
-                  "starting time":self.__startingTime,
-                  "ending time": self.__endingTime,
-                  "location":self.__location,
-                  "tasks": self.__tasks,
-                  "assignedVolunteers": self.__assignedVolunteers,
-                  "interest": self.__interest
-            }
-            
-            # getters & setters
-      def setTitle(self, title):
-            self.__title = title
-
-      def getTitle(self):
-            return self.__title
-      
-      def setDate(self,date):
-            self.__date = date
-
-      def getDate(self):
-            return self.__date
-      
-      def setStartingTime(self,time):
-            self.__startingTime = time
-
-      def getStartingTime(self):
-            return self.__startingTime
-      
-      def setEndTime(self,time):
-            self.__endingTime = time
-
-      def getEndTime(self):
-            return self.__endingTime
-      
-      def setLocation(self,location):
-            self.__location = location
-      
-      def getLocation(self):
-            return self.__location
-      
-      def setTasks(self,tasks):
-            self.__tasks = tasks
-      
-      def getTasks(self):
-            return self.__tasks
-      
-      def setAssignedVolunteers(self,volunteer):
-            self.__assignedVolunteers = volunteer
-
-      def getAssignedVolunteers(self):
-            return self.__assignedVolunteers
-      
-      def setInterest(self,interest):
-            self.__interest = interest
-      
-      def getInterest(self):
-            return self.__interest
-      
-
+      def getStatus(self):
+            return self.__status
+      #Exception Handling, task number not found!!!
             
       
-            
+      def __str__(self):
+        return f"Task No: {self.__taskNo}\nTitle: {self.__title}\nSkills: {self.__requiredSkills}\nCredit Hours: {self.__creditHour}\nVolunteers Needed: {self.__numOfVolunteersNeeded}\nStatus: {self.__status}\n===================="
+      
+
          
       
-class Organization:
-      """
-      """
-      organizationRecord = {}
-      def __init__(self) -> None:
-            self.__organization_name = str(input("Organization Name: ")).replace(" ","-")
-            self.__description = str(input("Description: "))
-            self.__opportunities = [ ]
-            self.__representatives = [ ]
-            
-            Organization.organizationRecord [self.__organization_name] = {
-                  "description": self.__description,
-                  "opportunities": self.__opportunities,
-                  "representatives": self.__representatives
-            }   
-            
-            # getters & setters
-      def setOrganizationName(self,name):
-            self.__organization_name = name
-      
-      def getOrganizationName(self):
-            return self.__organization_name
-      
-      def setDescription(self,desc):
-            self.__description = desc
-      
-      def getDescription(self):
-            return self.__description
-      
-      def setOpportunities(self,opportunities):
-            self.__opportunities = opportunities
-      
-      def getOpportunities(self):
-            return self.__opportunities
-      
-      def setRepresentatives(self,rep):
-            self.__representatives = rep
 
-      def getRepresentatives(self):
-            return self.__representatives
       
             
       
@@ -219,8 +89,10 @@ class Organization:
 class User:
       """
       Super class of three other classes (Administrator, Organization Representative, and Volunteer)
+      User class contains attributes : user id, full name, mobile number, email, education level, date of birth, and date of join.
       """
-      userRecord = {}
+      
+      userRecord = []
       
       today = datetime.now()
       year = str(today.year)
@@ -229,170 +101,402 @@ class User:
       strOfID = str(year + "0" + month + nums)
       randomID = int(strOfID)
       
-      def __init__(self, 
+      def __init__(self,
+                   fullname = '',mobile = '',email = '',educationLevel = '',DOJ = date, DOB = date
                    ) -> None:
             User.randomID += 1
-            self.__userID = User.randomID
-            print("Enter the following information")
-            self.__fullName = str(input("full name: "))
-            self.__mobile = str(input("mobile number: "))
-            self.__email = str(input("email: "))
-            self.__educationLevel = str(input("education level: "))
-            self.__DOB = input("date of birth (yyyy-mm-dd): ").split("-")
-            self.__DOB = date(int(self.__DOJ[0]), int(self.__DOJ[1]), int(self.__DOJ[2]))
+            self.__userID = User.randomID            
+            self.__fullName = fullname
+            self.__mobile = mobile
+            self.__email = email
+            self.__educationLevel = educationLevel
+            self.__DOB = DOB
+            self.__DOJ = DOJ
             self.__DOJ = datetime.today()
+            User.userRecord.append(self)
+
             
 
             # getters & setters
-      def setUserID(self,id):
+      def setUserID(self,
+                    id):
             self.__userID = id
       
       def getUserID(self):
             return self.__userID
       
-      def setFullName(self,name):
+      def setFullName(self,
+                      name):
             self.__fullName = name
       
       def getFullName(self):
-            return self.__fullName
+             return self.__fullName
+
       
-      def setMobile(self,mobile):
+      def setMobile(self,
+                    mobile):
             self.__mobile = mobile
       
       def getMobile(self):
             return self.__mobile
       
-      def setEmail(self,email):
+      def setEmail(self,
+                   email):
             self.__email = email
       
       def getEmail(self):
             return self.__email
 
-      def setEducationLevel(self,level):
+      def setEducationLevel(self,
+                            level):
             self.__educationLevel = level
       
       def getEducationLevel(self):
             return self.__educationLevel
       
-      def setDOB(self, date):
+      def setDOB(self,
+                 date):
             self.__DOB = date
       
       def getDOB(self):
             return self.__DOB
       
-      def setDOJ(self,date):
+      def setDOJ(self,
+                 date):
             self.__DOJ = date
       
       def getDOJ(self):
             return self.__DOJ
+      
+            """
+            Global functions 
+            """
+            
+      def __str__(self):
+            return f"User ID: {self.__userID}\nName: {self.__fullName}\nMobile: {self.__mobile}\nEamil: {self.__email}\nEducation Level: {self.__educationLevel}\nDate of Birth: {self.__DOB}\nDate of Join: {self.__DOJ}"
+
+
  
  
 
 
 class Volunteer(User):
       """
+      Child class of User with attributes : user id, skills, list of tasks, list of completed tasks, and total volunteering hours. 
       """
-      volunteerRecord = {}
-
+      volunteerRecord = []
+      
       def __init__(self,
-                   ) -> None:
-            super().__init__()
+                   fullname = '',mobile = '',email = '',educationLevel = '',DOJ = date, DOB = date, skills="") -> None:
+            super().__init__(fullname,mobile,email ,educationLevel,DOJ , DOB)
             self.__userID = "V" + str(self._User__userID)
-            self.__skills = input("skills separated by comma: ").split(",")
+            self.__skills = skills.split(",")
             self.__tasks = []
             self.__completedTasks = []
             self.__totalVolunteeringHrs = 0
-            
-            
-            Volunteer.volunteerRecord [self.__userID] = {
-                  "skills": self.__skills,
-                  "tasks": self.__tasks,
-                  "completedTasks": self.__completedTasks,
-                  "totalVolunteeringHrs" : self.__totalVolunteeringHrs
-               
-            }
-            
-            
-            User.userRecord [self.__userID] = {
-                  "fullname": self._User__fullName,
-                  "mobile": self._User__mobile,
-                  "email": self._User__email,
-                  "educationLevel": self._User__educationLevel,
-                  "DOJ": self._User__DOJ,
-                  "DOB": self._User__DOB      
-            }
+            Volunteer.volunteerRecord.append(self)
             
             # getters & setters
-      def setUserID(self,id):
+      def setUserID(self,
+                    id):
             self.__userID = id
       
       def getUserID(self):
             return self.__userID
       
-      def setSkills(self,skills):
-            self.__skills = skills
+      def setSkills(self,
+                    skills):
+            self.__skills = skills.split(",")
       
       def getSkills(self):
-            return self.__skills
+            data = ""
+            for skill in self.__skills:
+                  data += "\t" + str(skill) + "\n"
+            return data
       
-      def setTasks(self,tasks):
-            self.__tasks = tasks
+      
+      def addTask(self, task = Task()):
+               self.__tasks.append(task)
       
       def getTasks(self):
-            return self.__tasks
+            data = ""
+            for obj in self.__tasks:
+                  data += "\t" + str(obj) + "\n"
+            return data
       
-      def setCompletedTasks(self,compTasks):
-            self.__completedTasks = compTasks
+      
+      def setTotalVolunteerHours(self, 
+                                 sign = "", amount = 0):
+            if sign == "+": 
+                  self.__totalVolunteeringHrs += amount
+            elif sign == "-":
+                  self.__totalVolunteeringHrs -= amount
+            else: 
+                  return f"Error, check your inputs again"
+            
+            
+      def addCompletedTasks(self, task = Task()): # type: ignore
+            task.updateStatus("Completed")
+            self.setTotalVolunteerHours("+", task.getCreditHour())
+            self.__completedTasks.append(task)
+            self.__tasks.remove(task)
       
       def getCompletedTasks(self):
             return self.__completedTasks
       
-      def setTotalVolunteerHours(self, hrs):
-            self.__totalVolunteeringHrs = hrs
       
       def getTotalVolunteerHours(self):
             return self.__totalVolunteeringHrs
+      
+      
+            """
+            Global functions 
+            """
+            
+
+
+      def __str__(self):
+            return super().__str__()
 
             
+
+class Volunteer_Opportunity:
+      """
+      Opportunity class contains opportunities for volunteers. 
+      The class has attributes including title, date, starting time, ending time, location, list of tasks, list of assigned volunteers, and list of interested volunteers.
+      """
+      randomCounter = 0
+      
+      def __init__(self,
+                   title, date, startingTime, endingTime, location) -> None:
+            Volunteer_Opportunity.randomCounter += 1
+            self.__opportunityCode = Volunteer_Opportunity.randomCounter
+            self.__title = title
+            self.__date = date
+            self.__startingTime = startingTime
+            self.__endingTime = endingTime
+            self.__location = location
+            self.__tasks = []
+            self.__assignedVolunteers = []
+            self.__interest = []
+            
+            
+            # getters & setters
+      def setOpportunityCode(self,code):
+            self.__opportunityCode = code
+      
+      def getOpportunityCode(self):
+            return self.__opportunityCode
+      
+      def setTitle(self, 
+                   title):
+            self.__title = title
+
+      def getTitle(self):
+            return self.__title
+      
+      def setDate(self,
+                  date):
+            self.__date = date
+
+      def getDate(self):
+            return self.__date
+      
+      def setStartingTime(self,
+                          time):
+            self.__startingTime = time
+
+      def getStartingTime(self):
+            return self.__startingTime
+      
+      def setEndTime(self,
+                     time):
+            self.__endingTime = time
+
+      def getEndTime(self):
+            return self.__endingTime
+      
+      def setLocation(self,
+                      location):
+            self.__location = location
+      
+      def getLocation(self):
+            return self.__location
+      
+      
+      def addTask(self, 
+                  Task = Task()):
+            self.__tasks.append(Task)
+            
+      def getTasks(self):
+            return self.__tasks
+      
+      def addToAssignedVolunteers(self,
+                                volunteer = Volunteer()):
+            
+            if volunteer in self.__interest:
+                  self.__interest.remove(volunteer)
+                  self.__assignedVolunteers.append(volunteer)
+                  
+            for task in self.__tasks:
+                  volunteer.addTask(task)
+            
+      def getAssignedVolunteers(self):
+            data = ""
+            for obj in self.__assignedVolunteers:
+                  data += "\t" + {str(obj)} + "\n"
+            return data
+
+            
+            """
+            Global functions 
+            """
+            
+            
+      def setInterest(self,
+                      interest = Volunteer()):
+            self.__interest.append(interest)
+      
+      def getInterest(self):
+            data = ""
+            for obj in self.__interest:
+                  data += "\t" + str(obj) + "\n"
+            return data
+      
+
+            
+      def updateTaskStatus(self,
+                       task = Task(), status = ""):
+            if task in self.__tasks: 
+                  if status == "Completed": 
+                        for vol in self.__assignedVolunteers:
+                              vol.addCompletedTasks(task)
+                  else: 
+                        task.updateStatus(status)
+                        return f"task has been updated"
+            else:
+                  return f"{task}, is not assigned to this opportunity"
+                        
+    
+      def registerInterest(self, vol = Volunteer()):
+                  self.__interest.append(vol)
+                  
+             
+      def __str__(self):
+            return f"⎛ Title: {self.getTitle()} ⎞\n⎡ Date: {self.getDate()}\n├ Location: {self.getLocation()}\n├ Starting Time: {self.getStartingTime()}\n⎣ End Time: {self.getEndTime()}"
+        
+            
+            
+      
+                    
       
 
 class Organization_Representative(User):
       """
+      Child class of User, with attributes including user Id. 
       """
-      organizationRepRecord = {}
 
-
+      organizersRecord = []
       
-      def __init__(self) -> None:
-            super().__init__()
+      def __init__(self,
+                   fullname = '',mobile = '',email = '',educationLevel = '',DOJ = date, DOB = date
+                   ) -> None:
+            super().__init__(fullname,mobile,email ,educationLevel,DOJ , DOB)
             self.__userID = "O" + str(self._User__userID)
-      
-            Organization_Representative.organizationRepRecord [self.__userID] = {
-                  "User ID": self.__userID
-            }
-
-            User.userRecord [self.__userID] = {
-                  "fullname": self._User__fullName,
-                  "mobile": self._User__mobile,
-                  "email": self._User__email,
-                  "educationLevel": self._User__educationLevel,
-                  "DOJ": self._User__DOJ,
-                  "DOB": self._User__DOB      
-            }
+            
+            Organization_Representative.organizersRecord.append(self)
             
             # getters & setters
-      def setUserID(self,id):
-            self.__userID = id
       
       def getUserID(self):
             return self.__userID
       
 
+class Organization:
+      """
+      Organization class with attributes including organization name, description, list of opportunities, and list of representatives.
+      """
+      
+      organizationRecord = []
+      
+      def __init__(self,organization_name= " ",description ="", orgCode=0) -> None:
+            self.__organization_name = organization_name.replace(" ","-")
+            self.__organizationCode = "O" + str(orgCode)
+            self.__description = description
+            self.__opportunities = [ ]
+            self.__representatives = [ ]
+            
+            Organization.organizationRecord.append(self)
+            
+            # getters & setters
+      def setOrganizationName(self,
+                              name):
+            self.__organization_name = name
+      
+      def getOrganizationName(self):
+            return self.__organization_name
+      
+      def setDescription(self,
+                         desc):
+            self.__description = desc
+      
+      def getDescription(self):
+            return self.__description
+            
+      def createOpportunity(self, # update this function 
+                              title,date,startingTime,endingTime,location):
 
+            opportunity = Volunteer_Opportunity(title,date,startingTime,endingTime,location)
+            self.__opportunities.append(opportunity)
+            
+      def deleteOpportunity(self, opportunityCode):
+            try:
+                  for op in self.__opportunities:
+                        if op.getOpportunityCode() == opportunityCode:
+                              del(self)
+            except:
+                  return f"Failed to delete {self.__organization_name}"           
+        
+      
+      def getOpportunities(self):
+            data = "⎨Opportunities\n"
+            for obj in self.__opportunities:
+                  data += "\t" + str(obj) + "\n"
+            return data
+      
+      def setRepresentatives(self,
+                             rep = Organization_Representative()):
+            self.__representatives.append(rep)
 
+      def getRepresentatives(self):
+            data = ""
+            for op in self.__representatives:
+                  data += "\t" + str(op) + "\n"
+            return data
+      
+      def setOrgCode(self, code):
+                  self.__organizationCode = code
+      
+      def getOrgCode(self):
+            return self.__organizatioCode
+
+      
+      def addOpportunityToOrg(orgCode, otitle, odate, ostartingTime, oendingTime, olocation):
+            orgName = ""
+            
+            for org in Organization.organizationRecord: 
+                if org.getOrgCode() == orgCode: 
+                      orgName = org.getOrganizationName()
+                      org.createOpportunity(otitle, odate, ostartingTime, oendingTime, olocation)
+                else: 
+                      return f"Invalid Organization Code"
+            return f"Opportunity has been added to {orgName} successfully"
+                        
+            
 class Certificate:
       """
-      """
+      Certificate class with attributes including certificate ID, volunteer name, and issue date
+.      """
       today = datetime.now()
       year = str(today.year)
       nums = str("0000")
@@ -400,26 +504,22 @@ class Certificate:
       randomID = int(strOfID)      
       
       def __init__(self) -> None:
-            randomID += 1
-            self.__certificateID = randomID
-            self.__volunteer = "" # get volunteer name using volunteer ID
+            Certificate.randomID += 1
+            self.__certificateID = Certificate.randomID
+            self.__volunteer  =""
             self.__issueDate = datetime.today().date()  # printing the date of creation    
             
       # functions - getters & setters
-      def setCertificateID(self,id):
-            self.__certificateID = id
       
       def getCertificateID(self):
             return self.__certificateID
       
-      def setVolunteer(self,volunteer):
-            self.__volunteer = volunteer
+      def setVolunteer(self,
+                       volunteer = Volunteer()):
+            self.__volunteer = volunteer.getFullName()
       
       def getVolunteer(self):
             return self.__volunteer
-      
-      def setIssueDate(self,date):
-            self.__issueDate = date
       
       def getIssueDate(self):
             return self.__issueDate
@@ -429,37 +529,56 @@ class Certificate:
 
 class Administrator(User):
       """
-      """
-      administratorRecord = []
-      
-      def __init__(self)-> None:
-            super().__init__()
+      Child class of User with attributes including user 
+      """      
+      administratorsRecord = []
+     
+      def __init__(self,
+                   fullname = '',mobile = '',email = '',educationLevel = '',DOJ = date, DOB = date
+                   )-> None:
+            super().__init__(fullname, mobile, email, educationLevel, DOJ, DOB)
             self.__userID = "A" + str(self._User__userID)
 
-            Administrator.administratorRecord [self.__userID] = {
-                  "User ID": self.__userID
-            }
-            
-            User.userRecord [self.__userID] = {
-                  "fullname": self._User__fullName,
-                  "mobile": self._User__mobile,
-                  "email": self._User__email,
-                  "educationLevel": self._User__educationLevel,
-                  "DOJ": self._User__DOJ,
-                  "DOB": self._User__DOB      
-            }
 
             # getters & setters
-      def setUserID(self, id):
+      def setUserID(self, 
+                    id):
             self.__userID = id
       
       def getUserID(self):
             return self.__userID
+      
+      def registerOrganizations(self,name,code,description,opportunities,representatives):
+            try:
+                  org = Organization(name,code,description)
+                  return f"Organization registered"
+            except:
+                  return f"Failed to register organization"
+
+      def assignRepresentative(self, 
+                               org = Organization(), rep = Organization_Representative()):
+            if org:
+                  org.setRepresentatives(rep)
+            else:
+                  return f"Invalid Entry"
+            
+      def generateCertificate(self, 
+                              certificate = Certificate(), vol = Volunteer()):
+
+            certificate.setVolunteer(vol)
+            toPrint = f"====  -Certificate-  ====\n\tCID: {certificate.getCertificateID()}\n\tVolunteer: {self.getFullName()}\n\tIssue Date: {certificate.getIssueDate()}"
+            
+            return toPrint
+            
+                  
 
 
 
-#tests
-t = Task()
-e = Task()
-print(t.updateStatus(3, "completed"))
-print(Task.taskRecord)
+"""tasksRecord = []
+task1 = Task("Clean","Move",3,2)
+for task in tasksRecord:
+      print(task)
+task1.updateStatus(4,"Completed")
+print(task1.getStatus(4))
+print(task1.getNumOfVolunteerNeeded(4))"""
+
