+++ panel/functions.py
@@ -1,0 +1,1 @@
+from main.index import *
