+++ panel/system.py
@@ -1,4 +1,4 @@
-from index import *
+from functions import *
